All work must be done as the socrata user
Everything related to installing and configuring the gateway must be done as the
user Socrata will run as.  This user will always need a shell; however, the user
will only need sudo access during the initial installation.

Once the gateway is installed and connected, you can remove it's sudo access and
password.